Added a VS project file for tutorial30

+++ tutorial30/mesh.cpp
@@ -46,7 +46,7 @@ Mesh::MeshEntry::~MeshEntry()
 bool Mesh::MeshEntry::Init(const std::vector<Vertex>& Vertices,
                           const std::vector<unsigned int>& Indices)
 {
-    NumIndices = Indices.size();
+    NumIndices = (unsigned int)Indices.size();
 
     glGenBuffers(1, &VB);
   	glBindBuffer(GL_ARRAY_BUFFER, VB);
